+++ litho-it/src/test/java/com/facebook/litho/LayoutStateFutureReleaseTest.java
@@ -30,6 +30,7 @@ import android.os.Looper;
 import com.facebook.litho.config.ComponentsConfiguration;
 import com.facebook.litho.testing.Whitebox;
 import com.facebook.litho.testing.testrunner.LithoTestRunner;
+import com.facebook.rendercore.RunnableHandler;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
@@ -154,7 +155,7 @@ public class LayoutStateFutureReleaseTest {
     final Column column_0 = Column.create(mContext).child(new TestChildComponent()).build();
     final Column column = Column.create(mContext).child(child1).build();
 
-    LithoHandler handler =
+    RunnableHandler handler =
         ThreadPoolLayoutHandler.getNewInstance(new LayoutThreadPoolConfigurationImpl(1, 1, 5));
 
     componentTree = ComponentTree.create(mContext, column_0).layoutThreadHandler(handler).build();